Various metadata improvements
Added support for pulling metadata out of git (seperate metadata than
that of the github plugin). This data can be used to list first and last
edits, dates, authors, and total number of edits.

Added trial support for browser-sync on Windows machines (ONLY if it is
installed).

Fixed the last-modified meta tag

Added a bunch (of badly formatted) metadata to the container layout.
More work needed.

Standardized date formatting in other areas.