Changes to menu and collections
The 'correct' mindset for jekyll collections is to think of each as a
flat topic. Nesting folders and documents in a collection does not work
well in Jekyll. Rather Abstracting the grouping and ensuring that
collections are only one-level deep works better.

Thus the main menu is now defined by a menu.yml file which groups
together relevant collections to add in as headers.